Refactored terminal app into package.
