+++ run_gui.py
@@ -1,0 +1,9 @@
+import os
+from chem_data import data_set
+from chem_textual_ui.chem_ui import ChemApp
+
+
+def start_chem_app():
+    app = ChemApp()
+    app.run()
+start_chem_app()
