+++ terminal_version/term_app_helper_methods.py
